Fix hhvm test that broken with specialization
Summary:
This test broke with a seemingly-valid hacknative specialization pass
due to redeclaring a trait field with a different type. The specializer
cloned a function and refined it's This arg from BaseException to E,
at which point it resolved $message to the wrong type.

But the test doesn't typecheck...  After fixing the issues that
hh_single_typecheck complained about, including the redeclared
field types, it passes.

I'm assuming these changes are ok and the old test isn't really
valid hack. I couln't find any hg history indicating what the test
was really supposed to be testing... wdyt?

Reviewed By: alexeyt

Differential Revision: D16151883

fbshipit-source-id: 6d633bab2b9e6abe396c90632b2b497a4723b4d4