Option: allow specifying the naming table SQLite path via a --config override
Summary:
It has long been my belief that we should be adding settings we don't want users to see in our configs, so we can still set them via `--config` but not have to document them.

I wanted to be able to load the naming table easily from a previously saved version by pointing hh_server at a file of my choosing, hence this diff.

Reviewed By: pittsw

Differential Revision: D16113018

fbshipit-source-id: 8461680bc257878ae26178535b051e0afc48fb02