Do not mark ancestors of required classes as requirements
Summary:
In {D14631448}, we added tracking of requirements to the linearization, but the implementation has a bug. When we have a hierarchy like this:

```
interface I {}
class A implements I {}
trait T { require extends A; }
class B extends A { use T; }
```

The only requirement applied to B should be the one it received by using T, `require extends A`. However, the current implementation not only marks A as a requirement, it marks every *ancestor* of A as a requirement. As a result, shallow decl claims that B is required to extend/implement both A and I.

While the verifying of all these extra requirements does cause performance problems, they are often not a correctness problem, since B does in fact implement I (because it extends A). Correctness problems arise when a class implements an interface with an invariant type parameter twice with two different types (which is currently allowed by the typechecker; see T38737145). In this case, the interface requirements we aggregate can end up differing from the canonical instantiation of the doubly-implemented interface, causing shallow decl to emit errors where legacy decl does not (see the added test case require_extend_class_with_incompatible_instantiations_of_invariant_interface.php for an example of a situation where shallow decl emitted an error before this change).

The fix is simple. The original intent of the changed condition was to mark the mro_element representing `A` as a requirement when it is included in `T`'s linearization. Unfortunately, it was written carelessly--instead of marking only `A` as a requirement, it marks the *entire linearization of `A`*, so it marks `I` as a requirement as well.

We could fix this by changing the condition from `is_synthesized source` to `is_synthesized source && c.mro_name = class_name`, but there is no reason to do this check on every element in the linearization. We already use `Sequence.next` in `ancestor_linearization` to separate out the mro_element representing the required class (`A`) from the remainder of its linearization (`[I]`), so we might as well handle this logic specific to the required class after doing so.

Reviewed By: young-k

Differential Revision: D16111375

fbshipit-source-id: 509c3103be54295d4e1ccc6fad09b18826fb0d08