Include synthesized ancestors in the member resolution order of traits and interfaces only
Summary:
Synthesized ancestors (which are not marked as a requirement using `mro_required_at`) represent all ancestors of a class which was declared in a require-extends or require-implements clause. They are necessary for typechecking trait bodies and values of interface types.

For example, if we have:

```
class Base {
  public function foo(): void {}
}

class Child extends Base {}

trait T {
  require extends Child;
  public function testTrait() {
    $this->foo(); // legal
  }
}

interface I {
  require extends Child;
}

public function test_interface(I $x): void {
  $x->foo(); // legal
}
```

In order to typecheck `testTrait` and `test_interface`, we need `Base` to occur as a synthesized member in the linearizations of `T` and `I` so that we can look up the method `foo`:

```
Linearization of Base:
[Base]

Linearization of Child:
[Child, Base]

Linearization of T:
[T(copy_private_members), Child(requirement), Base(synthesized)]

Linearization of I:
[I, Child(requirement), Base(synthesized)]
```

Synthesized ancestors (which do not directly represent a requirement) are solely for this purpose, and are not used in other class kinds. However, we are currently aggregating them in the linearizations of concrete and abstract classes. There is no reason to do this--as long as we still aggregate requirements correctly, we can omit synthesized ancestors from the linearizations of concrete and abstract classes.

Reviewed By: young-k

Differential Revision: D16057838

fbshipit-source-id: f2ea0af93a83df9bf2e591a1e5f008dbd456e05f