Actually prevent users from extending and newing Closure.
Summary:
We tried to prevent users from extending Closure with arbitrary classes, but the code in the emitter and verifier that was doing this was doing case sensitive comparisons, so it didn't work. Fix these checks to be case insensitive, add unit tests.

We also tried to prevent users from directly constructing Closures by giving the Closure class a final constructor that raises an error, but constructors can be bypassed using reflection. Instead of using a constructor to enforce this, do what Awaitables do and make the instanceCtor raise an error, and have the CreateCl bytecode and c_Closure::clone bypass the instanceCtor to allocate them. This actually ensures that only CreateCl and Clone can produce a new Closure ObjectData. Add a unit test.

While we're here, delete a misleading/incorrect comment in ext_std_closure.h that mentions closureInstanceCtor, and stop marking the Closure class with __NativeData (since it doesn't actually use the native data machinery and has its own instanceCtor).

Reviewed By: markw65

Differential Revision: D16113186

fbshipit-source-id: 056e9ebc7168e87c8a3e7cbfadba522c57832381