Add support for const props on NativeData classes.
Summary:
Of all the classes with an instanceCtor, none of the isCppBuiltin ones (WaitHandles, closures, and collectons) have any user-accessible properties and none are user-extensible. The only non-isCppBuiltin classes with an instanceCtor are the ones with __NativeData; some of these are user-extensible.

Add support for const properties on classes with __NativeData by: storing pointers to 2 ctors in the class' extra data - one that constructs locked instances and one that constructs unlocked ones; splitting (templatizing on a bool) nativeDataInstanceCtor into these 2 versions, and using the version that constructs unlocked instances from code paths that start in the NewObj* bytecodes. Use the same ctor for both versions for the isCppBuiltin ones (since, as noted above, it doesn't matter).

Add a unit test with a user class with const props that extends a __NativeData class.

Reviewed By: markw65

Differential Revision: D16107454

fbshipit-source-id: eb0bf4a0d070433a0205ba3853c822f83b3c6101