Subtyping via where constraints
Summary:
[hack] Provide constraint information from class-level where clauses to sub typing.

This diff adds the handling of cases `C :< (this as I)` where the interface `I` has where clauses to mark its lower bounds and `I :< C` where `I` has upper bounds in its where clauses to subtyping.

For declaration such as `interface I where this as C` or `interface I where C super this`, the where constraint requires that any class implementing interface `I` must also extends the class `C`. This is enforced by checking the where constraint in declaration of those classes that the proposition `subtype (this as D) C` is true where `D` is any class trying to implement `I`. The proposition is simplified to `subtype D C` and is true only if `D` is a subclass of `C`. Besides, an object of type `this as I` is also eligible to be casted into some class `CParent` because of the constraint. This is equivalent to `subtype (this as I) CParent` being true. In this case, `subtype (this as I) CParent` is simplified to `subtype I CParent` in first place and then simplified to `subtype C CParent` where the class `C` appears in `I`'s where clauses in format of `this as C` or `C super this`. This check requires class `CParent` to be an actual superclass of `C`.

For declaration such as `interface I where C as this` or `interface I where this super C`, the constraint is, the interface `I` must and can only be implemented by the class `C` and class `C` must be a final class. This constraint is enforced by checking the proposition `subtype C (this as D)` which would fail except the `D` here is class `C`, which means `C` is the only class that can implement `I`. Besides, performing this check on declaration `class C implements I` requires `C` to be a final class.

Reviewed By: kmeht

Differential Revision: D16087019

fbshipit-source-id: d5db6441d18aa3e8602bfc92c724d05f4281194a