Avoid generating fresh type names in refinement in more cases
Summary:
When we refine to a generic type in `as` or `is` tests, we generate fresh type parameters with names such as `T#1`. If the generic type is covariant or contravariant, we lose no generality by instead substituting the upper (resp. lower) bound for the fresh type parameter name. For example:
```
class Cov<+Ti as I> { }
function do_refine(mixed $m):void {
  if ($m is Cov<_>) {
    // Here, we can treat $m as Cov<Ti#1> with Ti#1 <: I
    // But we lose nothing by refining to Cov<I>
  }
}
```
This is already what is done in `simplify_tpenv`. But in the case that there is no upper bound, we may as well use `mixed`, and for multiple upper bounds, we can now use an intersection. Dually, if there is no lower bound (for a contravariant type), we can use `nothing`, or a union of multiple lower bounds.

This leads to much simpler refined types in many cases.

Also in this diff: an access function `get_tpenv` for `env.lenv.tpenv`. A future diff will be moving this field into the per-continuation information, so we will change the implementation of `get_tpenv`.

Reviewed By: CatherineGasnier

Differential Revision: D16162036

fbshipit-source-id: 51a9fd5e20f87cb187c2dad397e1fcfba97cbf55