Make it an error to access a shape field known to not exist, even on the left of the coalesce operator
Summary: There is at the moment a discrepancy between typing of `$s['f']` on the left of the coalesce operator `??` and `Shapes::idx($s, 'f')`: the latter will produce an error if `$s` is a shape in which `'f'` is known to not exist, while the former will be accepted.  This is inconsistent because even internally in the type checker `$s['f'] ?? $d` is treated exactly like `Shapes::idx($s, 'f') ?? $d`.

Reviewed By: andrewjkennedy

Differential Revision: D16180713

fbshipit-source-id: 44d8eb5765e7b14ba76ec894c03a9fa9867eeb17