fix CLI client exit
Summary:
Per #8430, CLI client does not exit upon script termination because the
client's closure of the server's stderr does not trigger a read event
in all cases, leaving the client waiting to join a thread that will never
terminate. Introduce LogThread struct to encapsulate the thread and a
pipe used to signal the thread to exit (aka the self-pipe trick).

fixes facebook/hhvm#8430