Fix `hh --search-function` to filter by kind
Summary:
Some users have written scripts that rely on `hh --search-function x`.  This code previously would store the type value in `type_` and use a conversion function in HackSearchService to convert it from a string to a usable value.

The new symbol index pathways did not respect the `type_` parameter; and some of the work done to enforce "search-by-context" did not lend itself to "search-by-kind".

This diff adds a standard "string-to-kind" method that works for converting the strings defined in HackSearchService, and ensures that both Sqlite and Custom pathways can also search by kind as well as by context.

Reviewed By: pittsw

Differential Revision: D16185064

fbshipit-source-id: e0873989e0007565d72a13e72d924cfa5c326e68