Don't update search index if it's disabled
Summary:
Because the SQL-based naming table is not fast at O(repo) operations, it's affected by indexing worse than the memory-based naming table.

For trie-based indices, we index the symbols during the initialization sequence.

I noticed that even when we disable search indices by configuration, we still spend a good chunk of time indexing all the symbols.

This change is to address this issue.

Reviewed By: pittsw

Differential Revision: D16142578

fbshipit-source-id: 642185e9c3762cc811716bfae08c7e59ceb3286c