RFC: don't use Dfind in check mode
Summary:
I noticed that we wait upwards of 200 seconds for Dfind to be ready.
In check mode, we don't need a file watcher at all - we already get the sets of dirty/master files from Mercurial, so this is just a time penalty we don't need.

Most people don't run in check mode, so it should not affect them.
This affects saved states and people who test inits (me).

Reviewed By: pittsw

Differential Revision: D16143734

fbshipit-source-id: 35085f79e954a75591137f23e6777a338c281f7c