Fix bug in IdSet
Summary:
My previous refactoring had an oversight that caused the m_base of a
IdSet to always be set to 0 when adding the first value. This is just
an efficiency issue, but it was hiding another bug, where m_base
wasn't being set properly in the assignment operator. Fix both bugs.

Reviewed By: edwinsmith

Differential Revision: D16200308

fbshipit-source-id: 5ce6d5be1fc4256879dca1c590bf759b6344f213