Add typechecker option for lru_workers
Summary:
Add an option to gate whether `lru_workers` should be used instead of `MultiWorker`. Based the changes off of D16146597, not sure if everything is necessary in my use case!

I wanted to add it to `TypeCheckerOptions` to later do the lookup the same way it's done here: https://fburl.com/6ironuhx

As Ted mentioned, this option currently doesn't have any effects!

Differential Revision: D16204061

fbshipit-source-id: ae037394a3e9ab54da57eccf0d96947a7b56212a