Standardize namespace behavior - always delimit
Summary:
Autocomplete used a parameter, `delimit_on_namespaces`, which worked in either `true` or `false` mode.

Given the new work to track namespaces as their own entities, we are standardizing behavior to be predictable in all cases.  This means we can no longer support both modes.  This diff eliminates the flag and standardizes behaivor to produce the correct results.

Reviewed By: pittsw

Differential Revision: D16209698

fbshipit-source-id: 5a462c98fa9f778dbea277635d53567536240577