Add emitTLSLea()
Summary:
On some platforms (like x86), you can't use lea on a Vptr with a
segment register. Thus there's no easy way to turn the Vptr that
emitTLSAddr() returns into a Vreg. Add emitTLSLea() which does this
for you. On most platforms, this can just use a lea on the return of
emitTLSAddr(). On x86, we take advantage that the thread local base
points at a word containing its own address. Therefore a segmented
load and add of zero will obtain the address.

While we're at it, add support for segments to a few vasm instructions
which were missing it. This causes strange bugs where instructions are
encoded improperly.

Reviewed By: paulbiss

Differential Revision: D16215328

fbshipit-source-id: 54c18bf8679933ec0a63f8cfefb9dce3703f9048