Remote type checking options
Summary:
There are 5:

If set, distributes type checking to remote workers if the number of files to type check exceeds the threshold. If not set, then always checks everything locally.
  remote_type_check_threshold : int option;

If set, uses the key to fetch type checking jobs from XDB
  remote_worker_key : string option;

If set, uses the check ID when logging events in the context of remove init/work
  remote_check_id : string option;

Indicates the size of the job below which Eden should be used by the remote worker
  remote_worker_eden_checkout_threshold : int;

Dictates the number of remote type checking workers
  num_remote_workers : int;

Reviewed By: young-k

Differential Revision: D16146597

fbshipit-source-id: 66341ab38400ffac5de53d91e054d785c0709cc9