Add Autoload.Query to the repo-local options
Summary:
I'm adding a native Watchman-based autoloader to HHVM. Instead of giving HHVM a classmap or userland lambda, we now pass a Watchman query expression to HHVM. This query depends on the repo and needs to stay in sync with the repo, so the `.hhvmconfig.hdf` file in the repo seems like an obviously correct place to put it.

This is the first option which isn't related to the HackC parser, and it probably shouldn't fall into the `Parser {}` namespace. So instead of hacking it into the `Parser {}` namespace, I've created an `Autoload {}` namespace and a new `AUTOLOADFLAGS()` macro. That does make this otherwise trivial change into a nontrivial change, though.

Reviewed By: paulbiss

Differential Revision: D16193250

fbshipit-source-id: 0175c1824318bfb0e02f8375c332ff3baa2df4a3