Make invalid or sketchy truthiness tests lint warnings rather than Hack errors
Summary:
These are not "real" typing errors--they just indicate conventions which are likely to be misunderstood, or may behave in a way that wasn't intended by the author. We should just make them lint warnings instead.

In addition to making these lint warnings, this change also applies this linter to partial-mode files.

Reviewed By: hgoldstein

Differential Revision: D16225181

fbshipit-source-id: 2dab96ee7975954689767de3a8492d5e4dfdf25a