Explose naming table SQLite path
Summary: I want to use this path to determine whether to parallelize iteration over the naming table to MultiWorker (if it's SQLite-based) or to use the old way. Can also be used to decide whether to create a new naming table or to write incremental updates to an existing one.

Reviewed By: jewelpit

Differential Revision: D16216650

fbshipit-source-id: f911b9ac7f4b321cad0d410fe19f397f7ed2beaf