HHVM Debugger: Fix segfault in race where client removes a breakpoint while a request is loading a class
Summary:
There's a race if a request has started but not yet loaded a file with a breakpoint (or the breakpoint is unresovable like in a non-php file) and the breakpoint is removed from the client while the request is in onDefClass(). The request's list of unresolved breakpoints can end up containing a breakpoint ID for a breakpoint that no longer exists, so getBreakpointById returns nullptr.

This diff adds appropriate checks in that spot, but also defensive checks on all calls to getBreakpointById to ensure we're handling nullptr properly.

Reviewed By: alexeyt

Differential Revision: D16227981

fbshipit-source-id: 44591ba68cda726f9a6c3929dba5b13bab030648