Optimize lookup of expected tokens to O(1) time
Summary:
Instead of searching for relevant variants on the stack,
precompute the union as a bitmask and push it along the old info, i.e.,

    //     low bits          high bits
    <all_variant_bitmask> <last_variant_bit>

maintain the stack in the same fashion as so-called "Min Stack".
This way, the `Context::expects` method only needs to access the
top of the stack (which contains info about all expected tokens,
not just the current scope) as opposed to linearly searching.

Reviewed By: shiqicao

Differential Revision: D16266097

fbshipit-source-id: 7866bd83115668011482a4308c4908bda70c881d