Add <<__DisallowPHPArrays>> attribute
Summary:
This adds the `<<__DisallowPHPArrays>>` attribute to the typechecker. It will allow us to ban PHP arrays in type constants, which will allow them to be reified (see the thorough example in the task for why we need to do this).

This diff doesn't actually add the validation; it's only for the plumbing. The following diff will add validation and tests.

Reviewed By: oulgen

Differential Revision: D16077920

fbshipit-source-id: 487119adc7459553438fc730ee67188db29d54d3