hack/parser_rust: allow scanner inner loops to be specialized in the lexer
Summary:
Use `impl Fn` rather than `&dyn Fn` when passing predicates into
inner loops, as it allows them to be inlined and optimized. `&dyn` will cause
the call to be indirect via a vtable. (It's possible that llvm's
devirtualization optimization would fix this up, but better to be explicit.)

We can also pass the closure by value rather than reference, since nothing
seems to use it after the call.

This change converts everything to `impl Fn`. In theory this should have caused
lots of code-size bloat, but in practice it ended up being smaller than the original,
and ~.7% faster.

https://fburl.com/mei2r11t is a spreadsheet showing the variants I tested.

Reviewed By: shiqicao

Differential Revision: D16239098

fbshipit-source-id: b9db991500c556b93f590163c50b5ec9cfd26ca8