Make arraykey the default bound on enums
Summary: As a small step towards unification of abstract types, make *all* enums have an upper bound by interpreting the absence of `as t` by `as arraykey`. Unfortunately there are a couple of ways that this breaks: lint rules for equality, and type structure. Explicitly test for `arraykey` in these cases.

Reviewed By: manzyuk

Differential Revision: D16163835

fbshipit-source-id: e30a92ca0fd014ee012204253f9131af10fa609e