Avoid unnecessary intermediate vector allocations
Summary:
As in title / inline comments.

This might end up obsoleted by direct Ocaml syntax smart constructors, but I think it's an interesting example of how easy it is to waste a lot of time on temporary allocations.

I discovered it by breaking down the entire Rust -> OCaml conversion time in smaller logical steps. On the example inputs I tried the results were (middle one of three):

- entire program run: 42.6
- after removing OCaml heap allocations by no-op 33.7
- removing temporary vectors and replacing them by for loops: 31.0
- removing conversion completely: 30.1

That suggests that (after this fix) visiting the tree is pretty optimal, and the only cost of conversion are actual OCaml allocations.

Reviewed By: shiqicao

Differential Revision: D16271302

fbshipit-source-id: 8122a8f12e5ae2784856b64dc6f93ed2ce725ebf