Find docblocks within namespaces
Summary:
During debugging of serverless IDE, Waleed and I discovered that we often saw the message "Please rebase to find contains".  This is odd since "Str\contains" was a builtin, and shouldn't require rebasing.  The root cause was that the "Str\" was getting lost.

To address this, I preserve the fullname that includes the namespace in the LSP completion messages.

The end result is that docblocks for hhi builtins work again!

Note that this does not eliminate the usage of "file/line/column" docblocks, which existed previously.  File/line/column docblocks are needed to handle class methods.

Reviewed By: ljw1004

Differential Revision: D16271582

fbshipit-source-id: 2f969434d62c2722f836219f25a60203670107d9