Add gating logic to use lru_workers for typechecking
Summary:
Add the ability to use `lru_workers` instead of `MultiWorker` for typechecking.

In this version, workers are forked when the call is invoked, meaning that they have access to both the LRU in shared memory and the shared_memory used by MultiWorker.

Workers do not read/write values from the LRU cache yet, but they will very shortly (in a future diff :))

Reviewed By: jewelpit

Differential Revision: D16267846

fbshipit-source-id: cc555aab76f970d116b6a0084b245d729f91c05b