Disallow ... without type in function typehints
Summary:
For function typehints, we used to allow the variadic argument specifier without a typehint. This is an error in strict mode (and can be elevated to an error in partial mode as well). This makes it a parse error in the typechecker, requiring that we specify a typehint. A followup diff will make it a parse error in the compiler as well.

This will make implementation of variadic args for function type structures easier, which will then allow functions with variadic args to be reified type arguments.

Reviewed By: vassilmladenov

Differential Revision: D16231694

fbshipit-source-id: c444b3637c2a97a2c6c3d142c5e55fa7b985a240