Unban reification of dynamic and functions
Summary:
Recent changes to the runtime have allowed us to unban these types:
- `dynamic` now reifies differently from `mixed`
- Reification behavior and optimizations were implemented for functions.

We must still ban functions with variadic args for now, since that is not reflected in the type structure. I'm investigating the viability of banning the `...` syntax in Hack, so that function typehints must have a typehint for their variadic arg. That will make implementation of variadic args in type structures simpler.

Reviewed By: vassilmladenov

Differential Revision: D16229007

fbshipit-source-id: cc02c443b08704de5fa83fbbf2b2d5e80ffb59d3