Inout parameters should be treated invariantly for purposes of tyvars
Summary: Parameters marked `inout` should be treated as behaving invariantly (both covariant and contravariant) whereever variance is relevant. One such place is generation of fresh type variables, which are marked covariant, contravariant, or invariant, based on the position that they appear in the type of an expression, in particular the return type of a function that has been called. We solve covariant type variables to their lower bounds, and contravariant type variables to upper bounds, because such early solving does not reduce typeability of the surrounding context. But until now, we treated `inout` as contravariant, which leads to programs being rejected (see test). We should treat it as invariant.

Reviewed By: CatherineGasnier

Differential Revision: D16280760

fbshipit-source-id: db7cbe3a971b7bac25c09cf199fbdf0cd273f7d6