Remove namespace results from symbol search
Summary:
The server search was designed before namespaces were valid results.

Since namespaces do not have a location, namespace results come back with an invalid location.

Apparently, the current implementation of the workspace/symbol LSP message expects that results come back with a valid location.  When a namespace was returned in the result, LSP messages would crash.

This fixes the ServerSearch.go so that it will not crash on namespaces by doing the following:

* Add a new context, Ac_no_namespaces, which does not return namespace results.
* Switch ServerSearch.go to use the new context (previously it used "None")
* For safety' sake, alter the behavior of the "add_position_to_results" method so that it eliminates any results that lack a valid position.

Reviewed By: arxanas

Differential Revision: D16288844

fbshipit-source-id: c3f88c0071ad996aac1bf3b141bd53d4fd72bcd1