Adapt Class_get to class-level where clauses
Summary:
[hack] Enable static member lookup from class-level where clauses

This diff is a follow up of D16173682 in which we allow type checker to look up a member from upper bounds defined in class-level where clauses. The type of member found will be the intersection of results found in both the current class and its upper bounds.

In this diff, we extend this feature to static member as well. We'll have another diff to extend this to class const.

Reviewed By: CatherineGasnier

Differential Revision: D16228063

fbshipit-source-id: 704980fd6a2ce26ea57e6f78ea6502974c4f305e