Move rx_move_check out of nastCheck
Summary:
Moves the rx_move check out of nastCheck.

Allowed:
- assignment: `$a = HH\\Rx\\move($b);`
- call arguments: `foo(HH\\Rx\\move($b))`

Not Allowed:
- rx_move call inside rx_move call
  - There is another check that prevents this.
- rx_move used as an expression anywhere else

One of the differences between the tree operated on by nast_visitor and nastCheck is that in nast_visitor, identifiers are not elaborated. In nastCheck, all rx_move calls are elaborated to "\\HH\\Rx\\move". Therefore, in this check, before matching to the special name Sn.Rx.move, "\\" is prepended to the identifier.
See T47364994

Reviewed By: kmeht

Differential Revision: D15896578

fbshipit-source-id: 24e56272962e9d0f79bbfbbfeda77959d51df43f