Move scripts to opt-clang
Summary:
As per discussion in https://fb.workplace.com/groups/472626396119314/permalink/2623269061055026/ we propose to move all production builds to `mode/opt-clang`.

This diff updates all automated scripts to use the same build mode.

In testing, the performance of compilation in `mode/dev` and `mode/opt-clang` :
* HH build time mode/opt-clang: 866 seconds, 3,998MB raw, 305MB after removing debug symbols
* HH build time mode/dev: 816 seconds, 3,667MB raw, 1383MB after removing debug symbols.

The difference between 866 seconds and 816 seconds was not significant enough to override the benefits of re-using cached precompiled results.  If we keep all scripts on opt-clang, we expect to see better reuse of cached compiled binaries.  So let's switch the scripts that previously used mode/dev to mode/opt-clang as well.

Reviewed By: jewelpit

Differential Revision: D16345867

fbshipit-source-id: 2c2f7c2b564573e2cfb9eb1751ce2a83983dc9ed