Fix issues with ClsRefGetTS JIT implementation
Summary:
The ClsRefGetTS JIT implementation had a number of problems that meant
it always interp-one'd, even in the cases where it wasn't intended to
do so.

The first problem was that when checking for the presence of the
"generic_types" key in the input type-structure, it was using JmpZero
instead of JmpNZero, which meant we were calling the interpreter for
the *non-reified* case, instead of the reified case as intended.

This would have been a problem (because the rest of the implementation
only handles the non-reified case). However, when checking the type of
the "classname" field in the input type-structure, it was testing
!ts->type().maybe(TStr) and then punting if true. However, we should
be checking "clsName" here, not "ts" (which we've already established
is an array). Moreover, we should be checking maybe, not !maybe since
we want to punt if there's a chance the field is a string. Since "ts"
is never maybe a TStr, we'd always punt.

However, this whole check doesn't really do what's intended. In most
cases, we won't know the type of the field in the type-structure
(unless we have a constant or some HHBBC help). profiledArrayAccess
doesn't profile the field type, only its offset. So, in most cases,
we'll end up punting anyways. Moreover, since we punt (and not just
interpOne), we never actually emit the profiling code, so the
profiling does nothing, even in an optimized translation! The right
thing to do here is to just emit a runtime check of the field being
TStr and throw the exception if not, not check it statically and punt.

Fixing all this let us actually generate code for the non-reified
case, as intended, but it exposed another bug. Since this is the
non-reified case, we should store a Nullptr in the reified portion of
the class-ref slot (this is what the interpreter does). However the
JIT merely kills the reified portion of the slot, which means it can
contain arbitrary garbage. So, the next bytecode which attempts to
read the class-ref will crash. Actually store a Nullptr there (the JIT
will optimize away the store/load if it can).

Finally, one additional optimization. If we merely interpOne for the
reified case, we pessimize the non-reified case (which presumably is
more common). This is because after we return from the interpOne, the
JIT doesn't know which was stored in the class-ref slot, so we have to
reload everything at the phi point of the two cases. If instead we
perform a slow side-exit, we avoid generating a phi and the JIT can
(generally) optimize away the load/stores of the class-ref for the
non-reified case.

Reviewed By: oulgen

Differential Revision: D16355034

fbshipit-source-id: 057dac52f31d8994522adff74210e652c262f18c