Log precise hh_server startup time
Summary:
Andras Belokosztolszki asked me today how much time we exactly saved by improving the startup time of hh_server.

As it turns out, we don't log that information!  Bummer.  Let's log the "partially ready" and "fully ready" events so that we can track server average startup time and see whether we are improving it or degrading it.

Reviewed By: 2BitSalute, ljw1004

Differential Revision: D16361639

fbshipit-source-id: 39a184a24f05820ffbd82028ee13d9a8626ea1ed