Support coercion from abstract types to enforceable types
Summary:
Because we pessimize the upper bounds on erased generics, we end up with invalid subtypes in the case of the following:
```
function expect_int(int $i): void {}
function f<T as int>(T $t): void {
  // implicitly, T as ~int
  expect_int($t); // error, ~int </: int
}
```
`as` constraints on generics form an intersection. This implements coercion from an abstract type (such as erased an generic or abstract type constant) if an element the intersection that is its upper bound coerces to the target type, and the target type is enforceable.

Now we have two coercion primitives:

- `dynamic ~> T` if `T` is enforceable
- `Ta ~> T` if `T` is enforceable, `Ta <: Tb` and `Tb ~> T`

and one compound rule:

- `T1 | T2 ~> T` if `T1 ~> T` and `T2 ~> T`

Reviewed By: CatherineGasnier

Differential Revision: D15947100

fbshipit-source-id: 13fb16a4a3f1ddc7cc703cd4a16148106784def0