Allow top-level requires in strict mode files by default
Summary:
Files like this need manual refactoring:

```Hack
require_once('Bar.php');
class Foo extends Bar {}
```

`require_once()` can not be moved to an EntryPoint function in these cases, as it
`Bar` must be loadable before the runtime loads the class definition, which is
before the runtime definition.

More complex forms of this approach are found in > 1/3rd of the examples in the Hack
tutorial, so shipping this would require large amount of the Hack tutorial to be
moved to partial, and for partial mode to be documented.

Some possible approaches to unblock this:
- split it out to a separate error code to allow FIXME'ing - `/* HH_FIXME[1002] */` is
  extremely dangerous, and should probably be invalid. This would not unblock runtime
  changes. This is also wanted to allow redefining partial mode as "strict mode, but these error codes are suppressed"
- automatic autoloading (planned)
- change the runtime to allow the `require()` to be inside an entrypoint function - i.e.
  lazy-initialize the class definition (unlikely)

Reviewed By: kmeht

Differential Revision: D16382202

fbshipit-source-id: 5831d860c721e92817970c42f47078dea95341b1